[Journal] Additional cleanup of old journal files
