+++ components/Paper/forms/HubsSelector.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useCallback, useState } from 'react';
+import { useCallback } from 'react';
 import {
   MultiSelectOption,
   SearchableMultiSelect,
