+++ components/landing/HeroCardSwap.tsx
@@ -76,27 +76,6 @@ export function HeroCardSwap({ className = '' }: HeroCardSwapProps) {
               </video>
             </div>
           </Card>
-
-          <Card customClass="!bg-white !border-gray-200">
-            {/* Header Section */}
-            <div className="flex items-center space-x-3 px-6 py-4 bg-white border-b border-gray-100">
-              <Icon name="rhJournal2" size={20} color="black" />
-              <span className="text-xl font-semibold text-black">Publish</span>
-            </div>
-            {/* Video Section */}
-            <div className="p-4 flex-1 bg-gray-50 rounded-b-xl">
-              <video
-                className="w-full h-full object-cover rounded-xl shadow-lg"
-                autoPlay
-                loop
-                muted
-                playsInline
-              >
-                <source src="/videos/Publish_1080p.mp4" type="video/mp4" />
-                Your browser does not support the video tag.
-              </video>
-            </div>
-          </Card>
         </CardSwap>
       </div>
     </div>
