+++ public/ai.txt
@@ -23,7 +23,6 @@ Disallow: /notebook/
 Disallow: /moderators/
 Disallow: /expert-finder/
 Disallow: /settings
-Disallow: /paper/submit
 Disallow: /_next/
 Disallow: /sw.js
